intern -> alter-var-root+alter-meta!, untrace-var & trace-ns no-op when appropriate
Switch to

    (doto v
      (alter-var-root ...)
      (alter-meta! ...))

in trace-var and untrace-var for extra clarity.

Also make untrace-var a no-op for Vars which were not traced with
trace-var and make trace-ns a no-op for clojure.core and
clojure.tools.trace (since they include Vars used by the tracing
machinery itself, which should therefore not be traced; the user can
still trace Vars from these namespaces selectively with trace-var(s)).

Signed-off-by: Luc Préfontaine <lprefontaine@softaddicts.ca>